- Finalized possibility to copy on an existing instance - Added usage examples - Updated changelog
