+++ src/main/java/com/hotels/beans/transformer/Transformer.java
@@ -105,4 +105,9 @@ public interface Transformer {
      * @return the {@link Transformer} instance
      */
     Transformer setValidationDisabled(boolean validationDisabled);
+
+    /**
+     * Removes all the configured fields to skip.
+     */
+    void resetFieldsTransformationSkip();
 }
